Restructure base test with conftest
